add the organization name validation tooltip to the organization creation dialog
