+++ src/intl/en.json
@@ -244,6 +244,14 @@
         "description": "There are no logs to show, the latest logs may be older than 72 hours."
       }
     },
+    "organizationNameField": {
+      "maxLength": "Maximum 39 characters",
+      "letters": "Letters are only lowercase",
+      "startEndAlphanumeric": "Start and end with an alphanumeric symbol",
+      "whitespace": "No whitespace",
+      "alphanumeric": "No non-alphanumeric symbol except \"-\"",
+      "noConsecutiveDashes": "No consecutive dashes"
+    },
     "upgradeDialog": {
       "cardNumberLabel": "Card number",
       "expirationLabel": "Expiration",
@@ -1899,15 +1907,7 @@
         "title": "What is your organization's name?",
         "canBeChanged": "You can always change it later if you need to",
         "tooltip": "An organization is a space containing all your Koyeb resources that you can use for your own projects or to collaborate as a team. The organization name must be unique and is used to compose your apps public URL. You can change it at any time.",
-        "organizationNameLabel": "Organization's name",
-        "organizationNameRules": {
-          "maxLength": "Maximum 39 characters",
-          "letters": "Letters are only lowercase",
-          "startEndAlphanumeric": "Start and end with an alphanumeric symbol",
-          "whitespace": "No whitespace",
-          "alphanumeric": "No non-alphanumeric symbol except \"-\"",
-          "noConsecutiveDashes": "No consecutive dashes"
-        }
+        "organizationNameLabel": "Organization's name"
       },
       "qualification": {
         "title": "Tell us a bit about you",
